Add TicksLeft string
