+++ v1.4/Source/RimArchive/RimArchive/Thing/CompProperties_Healing.cs
@@ -33,7 +33,6 @@ public class CompHealing : ThingComp
     {
         base.Initialize(props);
         this._healsleft = Props.SingleUse ? 1 : Props.UseCount;
-        Log.Message($"{parent.def.defName} : {_healsleft} heals left");
     }
